python/ctpc-api/pyproject.toml
[project]
name = "ctpc-api"
version = "0.1.0"
description = "CT-PC FastAPI server for WinWerth automation"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pyautogui>=0.9.54",
    "pywinauto>=0.6.8",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "mss>=9.0.0",
    "pynput>=1.7.6",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.21.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
werth-autopilot = "WerthAutopilot:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
